split Websocket logic
